refactor(cli): issue + actions helpers を family modules へ移行 (SPEC-1942)
SPEC-1942 SC-025 進捗継続。 残 cli.rs 内の cache helper / gh wrapper を
該当 family へ移動。

cli/issue.rs へ移動 (pub(super) fn):
- load_or_refresh_issue / refresh_issue_cache
- load_or_refresh_linked_prs / linked_prs_cache_path
- read_linked_prs_cache / write_linked_prs_cache
- fetch_linked_prs_via_gh

cli/actions.rs へ移動 (pub(super) fn):
- fetch_actions_run_log_via_gh / fetch_actions_job_log_via_gh

cli.rs 1992 → 1752 行 (-240、 累計 -949 from 2701)。 env.rs 内の
super::fetch_linked_prs_via_gh / super::fetch_actions_*_via_gh 呼び出しは
super::issue::* / super::actions::* に更新。 cli.rs の test refs も
qualified path (crate::cli::issue::* / crate::cli::actions::*) に更新。

機能・契約・test 結果は不変 (lib 345 全 pass、 既知 flake test
app_runtime_viewport_and_geometry_updates_persist_workspace_state は
個別実行で pass、 broad concurrent run でのみ flake)。
